+++ functions/Widgets.fnc.php
@@ -261,10 +261,8 @@ function Widgets( $item, &$myextra = null )
 			' . _( 'Course' ) . '
 			</td><td>
 			<div id="course_div"></div> 
-			<a href="#" onclick=\'window.open(
-					"Modules.php?modname=misc/ChooseCourse.php",
-					"",
-					"scrollbars=yes,resizable=yes,width=800,height=400"
+			<a href="#" onclick=\'popups.open(
+					"Modules.php?modname=misc/ChooseCourse.php"
 				); return false;\'>' .
 				_( 'Choose' ) .
 			'</a>
@@ -319,10 +317,8 @@ function Widgets( $item, &$myextra = null )
 			'. _( 'Request' ) . '
 			</td><td>
 			<div id="request_div"></div> 
-			<a href="#" onclick=\'window.open(
-					"Modules.php?modname=misc/ChooseRequest.php",
-					"",
-					"scrollbars=yes,resizable=yes,width=800,height=400"
+			<a href="#" onclick=\'popups.open(
+					"Modules.php?modname=misc/ChooseRequest.php"
 				); return false;\'>' .
 				_( 'Choose' ) .
 			'</a>